# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.ArchiveListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.ArchiveResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handlers.ArchiveResponse:
    properties:
      artists:
        items:
          type: string
        type: array
      category:
        type: string
      characters:
        items:
          type: string
        type: array
      circles:
        items:
          type: string
        type: array
      created_at:
        type: string
      file_path:
        type: string
      id:
        type: string
      is_favorited:
        type: boolean
      language:
        type: string
      page_count:
        type: integer
      parodies:
        items:
          type: string
        type: array
      progress:
        $ref: '#/definitions/handlers.ProgressResponse'
      rating:
        type: integer
      release_date:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      thumbs_ready:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
    type: object
  handlers.ArtistListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.ArtistResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handlers.ArtistResponse:
    properties:
      aliases:
        items:
          type: string
        type: array
      count:
        type: integer
      id:
        type: integer
      name:
        type: string
      urls:
        items:
          type: string
        type: array
    type: object
...
  handlers.CreateArtistRequest:
    properties:
      aliases:
        items:
          type: string
        type: array
      name:
        type: string
      urls:
        items:
          type: string
        type: array
    type: object
...
  handlers.ProgressResponse:
    properties:
      completed:
        type: boolean
      current_page:
        type: integer
      last_read:
        type: string
    type: object
...
  handlers.UpdateArtistRequest:
    properties:
      aliases:
        description: |-
          Aliases/URLs: omit to leave unchanged, send [] to clear, or send a
          full list to replace it entirely.
        items:
          type: string
        type: array
      name:
        type: string
      urls:
        items:
          type: string
        type: array
    type: object
...
  response.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Self-hosted doujinshi library manager API
  title: Shoka API
  version: "1.0"
paths:
...
  /api/artists:
    get:
      parameters:
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ArtistListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: List artists with pagination
      tags:
      - artists
    post:
      consumes:
      - application/json
      parameters:
      - description: Artist to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateArtistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ArtistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Create an artist
      tags:
      - artists
  /api/artists/{id}:
    delete:
      description: Cascade-deletes the artist's aliases, URLs, and circle memberships,
        and unlinks it from every archive. Archive metadata/files are untouched.
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Delete an artist
      tags:
      - artists
    get:
      description: Includes the artist's aliases and URLs, unlike the list endpoints.
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ArtistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get a single artist
      tags:
      - artists
    patch:
      consumes:
      - application/json
      description: 'Partial update. Omit a field to leave it unchanged. For aliases/urls:
        omit to leave unchanged, send an empty array to clear, or send a full list
        to replace it — there''s no way to add/remove a single value.'
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateArtistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ArtistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update an artist
      tags:
      - artists
  /api/artists/{id}/archives:
    get:
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - default: 24
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ArchiveListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: List archives for an artist
      tags:
      - artists
  /api/artists/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ArtistResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: List every artist
      tags:
      - artists
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"